fix: countdown
