+++ src/components/dashboard/RecoveryInProgress/index.tsx
@@ -12,6 +12,7 @@ import { FEATURES } from '@/utils/chains'
 import { selectRecovery } from '@/store/recoverySlice'
 import type { RecoveryState } from '@/store/recoverySlice'
 import madProps from '@/utils/mad-props'
+import { getCountdown } from '@/utils/date'
 
 export function _RecoveryInProgress({
   blockTimestamp,
@@ -80,26 +81,12 @@ export function _RecoveryInProgress({
   )
 }
 
-export function _getCountdown(seconds: number): { days: number; hours: number; minutes: number } {
-  const MINUTE_IN_SECONDS = 60
-  const HOUR_IN_SECONDS = 60 * MINUTE_IN_SECONDS
-  const DAY_IN_SECONDS = 24 * HOUR_IN_SECONDS
-
-  const days = Math.floor(seconds / DAY_IN_SECONDS)
-
-  const remainingSeconds = seconds % DAY_IN_SECONDS
-  const hours = Math.floor(remainingSeconds / HOUR_IN_SECONDS)
-  const minutes = Math.floor((remainingSeconds % HOUR_IN_SECONDS) / MINUTE_IN_SECONDS)
-
-  return { days, hours, minutes }
-}
-
 function Countdown({ seconds }: { seconds: number }): ReactElement | null {
   if (seconds <= 0) {
     return null
   }
 
-  const { days, hours, minutes } = _getCountdown(seconds)
+  const { days, hours, minutes } = getCountdown(seconds)
 
   return (
     <Box display="flex" gap={1}>
